chore: tweak
- Separate src/index.ts
- Change API (`wakatsu` -> `wakati`)
- Fix tests
- Update README.md
- Update dependencies
