+++ test/index.mjs
@@ -3,20 +3,20 @@ import {
 	analyzeSync,
 	tokenize,
 	tokenizeSync,
-	wakatsu,
-	wakatsuSync,
+	wakati,
+	wakatiSync,
 } from '@enjoyjs/node-mecab';
 
 const text = 'こんにちは世界';
 
-Promise.all([analyze(text), tokenize(text), wakatsu(text)]).then((results) => {
+Promise.all([analyze(text), tokenize(text), wakati(text)]).then((results) => {
 	for (const result of results) {
-		console.log(result);
+		console.dir(result, {depth: null});
 	}
 });
 
-const syncs = [analyzeSync, tokenizeSync, wakatsuSync];
+const syncs = [analyzeSync, tokenizeSync, wakatiSync];
 
 for (const sync of syncs) {
-	console.log(sync(text));
+	console.dir(sync(text), {depth: null});
 }
